# Non-uniform boundary conditions