This pull request adds the capability of using space and time-varying inflow boundary conditions as well as prescribing space-time body force to the momentum step.

The major change is that now the `u_BC::Function` passed to the `Flow` is of the form `u_BC(i,x,t)` where `i` is the component, `x` the location in physical space and `t` the time. The flow is then accelerated correctly, and the boundary conditions on the velocity field are applied correctly. This allows to prescribe velocity profiles at the inlet and outlet. 

Another feature grabbed from #174 is to enable passing either a `body_force::AbstractArray` or `body_force::Function=(i,x,t)->...` to `mom_step!` and `sim_step!`, which allows the user to either prescribe some custom term or accelerate the flow locally.