@marinlauber 2025-02-27
Testing the allocation issue of this PR.

```julia
using WaterLily,StaticArrays,ForwardDiff
import WaterLily: accelerate!,@loop,size_u,slice
using BenchmarkTools

# classical accelerate!
# accelerate!(r,dt,g::Function,::Tuple,t=sum(dt)) = for i ∈ 1:last(size(r))
    # r[..,i] .+= g(i,t)
# end

accelerate_loop!(r,dt,g::Function,::Tuple,t=sum(dt)) = for i ∈ 1:last(size(r))
    @loop r[I,i] += g(i,t) over I ∈ CartesianIndices(Base.front(size(r)))
end

new_accelerate!(r,g::Function,t) = for i ∈ 1:last(size(r))
    @loop r[I,i] += g(i,loc(i,I,eltype(r)),t) over I ∈ CartesianIndices(Base.front(size(r)))
end

# some array to test this
N = 2^7
u = zeros(Float32,N,N,N,3);

# old method
@btime accelerate!(u,[1.0],(i,t)->sin(2π*t)*cos(2π*t)*cos(2π*t),()) # 5.265 ms (12 allocations: 24.00 MiB)
# same but loop
@btime accelerate_loop!(u,[1.0],(i,t)->sin(2π*t)*cos(2π*t)*cos(2π*t),()) # 9.490 ms (852 allocations: 67.02 KiB)
# new method
@btime new_accelerate!(u,(i,x,t)->sin(2π*t)*cos(2π*t)*cos(2π*t),1.0) # 9.516 ms (849 allocations: 66.94 KiB)
@btime new_accelerate!(u,(i,x,t)->sin(2π*x[1])*cos(2π*x[2])*cos(2π*x[3]),1.0) # 10.829 ms (849 allocations: 66.94 KiB)

# the new BC function
new_BC!(a,U,saveexit=false,perdir=(),t=0) = new_BC!(a,(i,x,t)->U[i],saveexit,perdir,t)
function new_BC!(a,uBC::Function,saveexit=false,perdir=(),t=0)
    N,n = size_u(a)
    for i ∈ 1:n, j ∈ 1:n
        if j in perdir
            @loop a[I,i] = a[CIj(j,I,N[j]-1),i] over I ∈ slice(N,1,j)
            @loop a[I,i] = a[CIj(j,I,2),i] over I ∈ slice(N,N[j],j)
        else
            if i==j # Normal direction, Dirichlet
                for s ∈ (1,2)
                    @loop a[I,i] = uBC(i,loc(i,I),t) over I ∈ slice(N,s,j)
                end
                (!saveexit || i>1) && (@loop a[I,i] = uBC(i,loc(i,I),t) over I ∈ slice(N,N[j],j)) # overwrite exit
            else    # Tangential directions, Neumann
                @loop a[I,i] = a[I+δ(j,I),i] over I ∈ slice(N,1,j)
                @loop a[I,i] = a[I-δ(j,I),i] over I ∈ slice(N,N[j],j)
            end
        end
    end
end

# very simple BCs
U = SA[1.0,0.,0.]
Ubc(i,I,t) = i==1 ? 1 : 0

# test the bcs
@btime BC!(u,U,false,()) # 1.025 ms μs (5905 allocations: 477.30 KiB)
# new function with an array
@btime new_BC!(u,U,false,(),0) # 1.038 ms (5905 allocations: 477.30 KiB)
# new function with a function
@btime new_BC!(u,Ubc,false,(),0) # 770.551 μs (5878 allocations: 468.30 KiB)
@btime Ubc(1,SA[10,10,10],0) # 1.031 ns (0 allocations: 0 bytes)

# more complex function
function u_pipe(i,x,t)
    i ≠ 1 && return 0
    r = √sum(abs2,SA[x[2],x[3]].-N÷2)
    return 2r>N-3 ? 0 : 1-r^2/N^2 
end
@btime new_BC!(u,u_pipe,false,(),0) # 2.358 ms (@TODO 644854 allocations: 10.96 MiB)
@btime u_pipe(1,SA[10,10,10],0)  # this is what is killing us actually
```