@marinlauber 2025-02-27
In summary, more allocations are in the `@loop` version of `accelerate!`, regardless of the function arguments. This is related to having to change from
```julia
r[..,i] .+= g(i,t)
```
to 
```julia
@loop r[I,i] += g(i,t) over I ∈ CartesianIndices(Base.front(size(r)))
```
but this is negligible compared to the BCs.