@marinlauber 2025-02-27
What kills the allocation is when the BCs are applied to the field. if the uBC function allocates, when the BC is applied, the function massively allocates.

```julia
# more complex function
function u_pipe(i,x,t)
    i ≠ 1 && return 0
    r = √sum(abs2,SA[x[2],x[3]].-N÷2)
    return 2r>N-3 ? 0 : 1-r^2/N^2 
end
@btime new_BC!($u,$u_pipe,$conv_exit,$tu,$t) # 2.358 ms (644854allocations: 10.96 MiB)
@btime u_pipe($i,$x,$t)  # 139.873 ns (9 allocations: 144 bytes) this is what is killing us
```