@marinlauber 2025-02-27
The allocation issue is solved (see below), and I have included stuff from #174, since it is just a variation of the general body force case `g(i,x,t)`

I have also tested the new interface.

```julia
     Testing Running tests...
[ Info: Test backends: Array, CuArray
Test Summary: | Pass  Total   Time
util.jl       |   44     44  34.8s
Test Summary: | Pass  Total   Time
Poisson.jl    |   14     14  29.9s
Test Summary:        | Pass  Total   Time
MultiLevelPoisson.jl |   13     13  12.0s
Test Summary: | Pass  Total   Time
Flow.jl       |   30     30  24.2s
Test Summary: | Pass  Total  Time
Body.jl       |    3      3  0.1s
Test Summary: | Pass  Total   Time
AutoBody.jl   |   16     16  13.7s
Test Summary:        | Pass  Total  Time
Flow.jl periodic TGV |    2      2  8.2s
Test Summary: | Pass  Total   Time
ForwardDiff   |    2      2  44.3s
Test Summary:                      | Pass  Total   Time
Flow.jl with increasing body force |    2      2  13.1s
Test Summary: | Pass  Total   Time
Metrics.jl    |   37     37  20.1s
Test Summary: | Pass  Total  Time
WaterLily.jl  |   30     30  8.8s
Test Summary: | Pass  Total   Time
VTKExt.jl     |   28     28  23.5s
     Testing WaterLily tests passed 
```
The results of `allocation.jl` are here.

```julia
▶ Allocated 6 KiB
▶ Allocated 18 KiB
Test Summary:         | Pass  Total   Time
mom_step! allocations |    2      2  34.6s
     Testing WaterLily tests passed 
```