@marinlauber 2025-02-27
commit `633b6d9` passes all the tests and `allocation.jl`

```julia
     Testing Running tests...
[ Info: Test backends: Array, CuArray
Test Summary: | Pass  Total   Time
util.jl       |   44     44  34.8s
Test Summary: | Pass  Total   Time
Poisson.jl    |   14     14  30.0s
Test Summary:        | Pass  Total   Time
MultiLevelPoisson.jl |   13     13  11.9s
Test Summary: | Pass  Total   Time
Flow.jl       |   30     30  25.4s
Test Summary: | Pass  Total  Time
Body.jl       |    3      3  0.1s
Test Summary: | Pass  Total   Time
AutoBody.jl   |   16     16  14.2s
Test Summary:        | Pass  Total  Time
Flow.jl periodic TGV |    2      2  8.3s
Test Summary: | Pass  Total   Time
ForwardDiff   |    2      2  44.5s
Test Summary:                      | Pass  Total   Time
Flow.jl with increasing body force |    2      2  13.2s
Test Summary: | Pass  Total   Time
Metrics.jl    |   37     37  20.3s
Test Summary: | Pass  Total  Time
WaterLily.jl  |   30     30  8.8s
Test Summary: | Pass  Total   Time
VTKExt.jl     |   28     28  24.3s
     Testing WaterLily tests passed 


```


```julia
▶ Allocated 6 KiB
▶ Allocated 16 KiB
Test Summary:         | Pass  Total   Time
mom_step! allocations |    2      2  35.0s
     Testing WaterLily tests passed 
```