@marinlauber 2025-02-27
@weymouth @b-fg let me know what you think. It seems that we could maybe clean a bit more the `acceleration!` / `body_force!` interface.