@b-fg 2025-03-03
Just had a quick look and I like it! Also, I guess this PR solves https://github.com/WaterLily-jl/WaterLily.jl/pull/174 too, right? Happy to close that one if so. Also I like this `body_force` implementation, which can be array or function. It will be useful for including turbulence models ;)