@marinlauber 2025-03-03
Yes, this fixes #174 (the body_force part of it). 

There is still one part of the test fails, which is related to constructing the simulation with a `uBC(i,t)`, and internally transforming that into a `uBC(i,x,t)`. The behaviour seems to be breaking the impulsive cylinder force test.

I am not sure if we want to keep the possibility of constructing the sim with three different velocity BC, `Tuple`, Function{Int,Number}` and `Function{Int,SVector,Number}` or if we change everything to the second function type.