@marinlauber 2025-03-03
> Just had a quick look and I like it! Also, I guess this PR solves #174 too, right? Happy to close that one if so. Also I like this `body_force` implementation, which can be array or function. It will be useful for including turbulence models ;)

That was the idea: simple body force can be implemented as g(i,x,t), but more complex like turbulence models need to access `Flow` variables, which not easy to implement just with a function ;)